@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, FormControl, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, UntypedFormControl, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 
 import { Errors, UserService } from '../core';
@@ -13,13 +13,13 @@ export class AuthComponent implements OnInit {
   title: String = '';
   errors: Errors = {errors: {}};
   isSubmitting = false;
-  authForm: FormGroup;
+  authForm: UntypedFormGroup;
 
   constructor(
     private route: ActivatedRoute,
     private router: Router,
     private userService: UserService,
-    private fb: FormBuilder
+    private fb: UntypedFormBuilder
   ) {
     // use FormBuilder to create a form group
     this.authForm = this.fb.group({
@@ -36,7 +36,7 @@ export class AuthComponent implements OnInit {
       this.title = (this.authType === 'login') ? 'Sign in' : 'Sign up';
       // add form control for username if this is the register page
       if (this.authType === 'register') {
-        this.authForm.addControl('username', new FormControl());
+        this.authForm.addControl('username', new UntypedFormControl());
       }
     });
   }
